requirements.txt
anthropic>=0.25.0
google-cloud-dialogflow-cx>=1.30.0
google-cloud-texttospeech>=2.16.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.32.0
groq>=0.5.0
rank_bm25>=0.2.2
rapidfuzz>=3.6.0
httpx>=0.27.0
python-dateutil>=2.9.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pinecone>=6.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
# Optional TTS:
# gTTS>=2.5.0
# Optional PDF ingestion:
# pdfplumber>=0.11.0
